Revert "rename assets -> public"
