+++ site/.vitepress/config.mjs
@@ -24,14 +24,12 @@ export default defineConfig({
   outDir: "../public",
   head: [
     ['link', { rel: 'stylesheet', href: 'https://unpkg.com/tailwindcss@2.0.4/dist/tailwind.min.css' }],
-    ['link', { rel: "icon", sizes: "16x16", type: "image/png", href: "https://staging.lightningjs.io/assets/favicons/lng_16x16.png"}],
+    ['link', { rel: "icon", sizes: "16x16", type: "image/png", href: "/assets/favicons/lng_16x16.png"}],
     ['link', { rel: "icon", sizes: "32x32", type: "image/png", href: "/assets/favicons/lng_32x32.png"}]
   ],
   transformPageData: (pageData) => {
     const conicalPath = pageData.relativePath.replace(/index\.md$/, '').replace(/\.md$/, '.html');
     const {title, description} = pageData.frontmatter;
-
-    console.log('test', import.meta)
     pageData.frontmatter.head ??= [];
     pageData.frontmatter.head.push(['meta', {name: 'og:url', content: new URL(conicalPath, import.meta.url)}])
     pageData.frontmatter.head.push(['meta', {name: 'og:title', content: pageData.frontmatter.layout === 'home2' ? 'Lightningjs' : `${title} | Lightningjs`}])
